Add quantity to ComponentProductType
Add a quantity to ComponentProductType, which allows the absolute quantity of a product used in a mix to be specified. This is in addition to the `percent` wherever possible, but supports cases where it is too hard to calculate percent (e.g., dissolving granules in a solution without knowing the density of the granules).
Resolves #62